+++ UfcAPI/Models/Finance.cs
@@ -6,7 +6,7 @@ namespace UfcAPI.Models;
 public class Finance : IFinance
 {
   public  int Id { get; set; }
-  public  int MoneyLeft { get; set; }
+  public  double MoneyLeft { get; set; }
    public int NumberOfPurchases { get; set; }
-   public int MoneySpent { get; set; }
+   public double MoneySpent { get; set; }
 }
